% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_estout.R
\name{fix_estout}
\alias{fix_estout}
\title{Convert STATA estout .csv to R data frames}
\usage{
fix_estout(fn)
}
\arguments{
\item{fn}{A full (character) filename path pointing to the desired .csv
created by estout.}
}
\value{
A list will be returned containing two data frames:
  \code{reg.results} giving the point estimates of the regression output and
  \code{reg.ses} giving the second line of the outputs (i.e. SEs or T values)
}
\description{
Get point estimates and SEs/T-values as a numeric data frame
  without the extra characters or formatting issues inputted by estout.
}
\details{
Ever wanted to just access the numbers from estout without
  having to deal with all the extra fluff included to make the output 
  readable by Excel, Word, or LaTeX? \code{fix_estout} is for you!
  \code{fix_estout} takes in a filename path to a .csv created by STATA's
  estout, and returns two data frames (in a list): \code{reg.results} giving
  the point estimates and \code{reg.ses} giving whatever second line option
  was picked (i.e. standard errors or t-values). All fluff ("=","()", etc.)
  is removed. The regresands are the column names (with .1, .2, as R does by
  default, for multiple specifications - a future version may offer the
  option to put different model specifications in different list elements),
  the regressors are the row names. The actual data frame is therefore purely
  numeric.
}
\examples{
\dontrun{fix_estout("~/code/regression_output.csv")}
}
