\name{various.utilities-package}
\alias{various.utilities-package}
\alias{various.utilities}
\docType{package}
\title{
\packageTitle{various.utilities}
}
\description{
\packageDescription{various.utilities}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{various.utilities}
\packageIndices{various.utilities}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{various.utilities}

Maintainer: \packageMaintainer{various.utilities}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
